requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
openai>=2.24.0
azure-ai-formrecognizer==3.3.3
azure-identity==1.17.1
pydantic==2.9.2
gunicorn==22.0.0
